♻️Refactor: Replace setters with factory methods, Use `BigDecimal` for monetary field
- 엔티티 클래스에서 @Setter 를 제거하고, 객체 생성 시 정적 팩토리 메서드를 활용
- 금액 관련 필드 타입을 BigDecimal 로 변경

Resolved #43, #44

+++ src/main/java/com/jane/ecommerce/interfaces/dto/item/ItemResponse.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -12,7 +14,7 @@ public class ItemResponse {
 
     String itemId;
     String name;
-    Long price;
+    BigDecimal price;
     int stock;
 
     public static ItemResponse from(Item item) {
