+++ src/main/java/com/jane/ecommerce/interfaces/dto/user/BalanceResponse.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 public class BalanceResponse {
 
     Long userId;
-    Long balance;
+    BigDecimal balance;
 }
